Added auth methods in Service

+++ MondialMunch/MondialMunchService.cs
@@ -1,7 +1,16 @@
 namespace MondialMunch;
 
 public class MondialMunchService {
+    private User? _current_user;
     private readonly MondialMunchContext _context;
+
+    public User? CurrentUser {
+        get => this._current_user;
+        set {
+            if (value != null && !value.IsAuthenticated) throw new Exception("Current user must be authenticated.");
+            this._current_user = value;
+        }
+    }
 
     public MondialMunchService(MondialMunchContext context) {
         _context = context;
@@ -12,6 +21,7 @@ public class MondialMunchService {
     public List<Country> GetCountries() => _context.Countries.ToList();
     public List<DietaryTag> GetDietaryTags() => _context.DietaryTags.ToList();
 
+    public User? GetUserByUsername(string name) => _context.Users.FirstOrDefault(user => user.Name == name);
     public Country? GetCountryByName(string name) => _context.Countries.FirstOrDefault(country => country.Name == name);
 
     public void AddRecipe(Recipe recipe) {
@@ -20,16 +30,40 @@ public class MondialMunchService {
     }
 
     public void AddUser(User user) {
+        if (!user.IsAuthenticated) throw new Exception("User objects must be authenticated before entering database.");
+
         _context.Users.Add(user);
         _context.SaveChanges();
     }
 
+    public void FavoriteRecipe(Recipe recipe) {
+        if (_current_user == null) throw new Exception("Not logged in.");
+        if (_current_user.FavouriteRecipies.Contains(recipe)) throw new Exception("Recipe is already in favorites.");
+
+        _current_user.FavouriteRecipies.Add(recipe);
+        _context.SaveChanges();
+    }
+
+    public void UnfavoriteRecipe(Recipe recipe) {
+        if (_current_user == null) throw new Exception("Not logged in.");
+        if (!_current_user.FavouriteRecipies.Contains(recipe)) throw new Exception("Recipe is not in favorites.");
+
+        _current_user.FavouriteRecipies.Remove(recipe);
+        _context.SaveChanges();
+    }
+
     public void DeleteRecipe(Recipe recipe) {
+        if (_current_user == null) throw new Exception("Not logged in.");
+        if (_current_user.Name != recipe.Creator.Name) throw new Exception("You do not own this recipe.");
+
         _context.Recipes.Remove(recipe);
         _context.SaveChanges();
     }
 
     public void DeleteUser(User user) {
+        if (_current_user == null) throw new Exception("Not logged in.");
+        if (_current_user.Name != user.Name) throw new Exception("You cannot delete this user.");
+
         _context.Users.Remove(user);
         _context.SaveChanges();
     }
